Fix deployment files to unblock spec
The format parameter needs to not have a leading . or all of the spec
tests need to be fixed to pass it in. As removing it is the simpler
option this does that.

Secondly, the calculated file names for the deploy need to have the file
extension . added. Also, the calculated file name for pre-7.0.0 had been
destroyed, this puts it back to what it should be.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>