Changes for java-11 and JIRA >8.12 compatibility
As of openjdk 11 it is necessary to make the extra
opts manageable so that -XX:+PrintGCDateStamps can
be replaced with -Xlog:::time,uptime,level,tags.
This change introduces the new variable into
init.ppp and uses it for the JVM_EXTRA_ARGS variable
in setenv.sh.

Further with the new mysql-connector (as of version) 8,
the archive does not contain a "-bin" jarfile. Instead
the connector is just named mysql-connector-java-<version>.jar

Updating setenv.sh to JIRA 8.12.1 options and replace
the old scope.lookupvar usage.

Move jvm version dependent configuration to setenv.sh.
Either specify jvm_type = custom and use the jvm_* parameters
or use the predefined values in the setenv.sh by using
jvm_type =  'openjdk-11' / 'oracle-jdk-1.8'